+++ CHANGELOG.md
@@ -19,6 +19,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
+## [0.3.3] - 2026-04-15
+
+### Added
+
+ - New provider admin database models for the FGA database provider.
+
+### Changed
+
+ - Refactored provider admin code in the FGA validator and database provider.
+
+### Fixed
+
+ - Small fix to correct a failing test.
+
+
 ## [0.3.1] - 2026-03-17
 
 ### Added